requirements.txt
watchdog>=3.0.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0
playwright>=1.40.0
apscheduler>=3.10.0
requests>=2.31.0
python-dotenv>=1.0.0
tweepy>=4.14.0
facebook-sdk>=3.1.0
retry>=0.9.2
